refactoring - sideMenu component in commons
