+++ packages/pn-commons/src/test-utils.tsx
@@ -1,0 +1,19 @@
+import { render, RenderOptions } from '@testing-library/react';
+import { ReactElement, ReactNode } from 'react';
+import { BrowserRouter } from 'react-router-dom';
+
+const AllTheProviders = ({children}: {children: ReactNode}) => {
+  return (
+    <BrowserRouter>
+        {children}
+    </BrowserRouter>
+  )
+}
+
+const customRender = (
+  ui: ReactElement,
+  options?: Omit<RenderOptions, 'wrapper'>,
+) => render(ui, {wrapper: AllTheProviders, ...options})
+
+export * from '@testing-library/react'
+export {customRender as render}
