+++ packages/pn-commons/src/types/SideMenuItem.ts
@@ -4,4 +4,5 @@ export interface SideMenuItem {
   label: string;
   icon: SvgIconComponent;
   route: string;
+  children?: Array<SideMenuItem>;
 }
