--- packages/pn-personafisica-webapp/src/component/SideMenu/SideMenuItem.ts
@@ -1,7 +1,0 @@
-import { SvgIconComponent } from '@mui/icons-material';
-
-export interface SideMenuItem {
-  label: string;
-  icon: SvgIconComponent;
-  route: string;
-}
